+++ Sources/KeyboardKit/Gestures/KeyboardGesture.swift
@@ -14,22 +14,22 @@ import Foundation
  */
 public enum KeyboardGesture: String, CaseIterable, Codable, Equatable, Identifiable {
 
-    /// Occurs when a button is pressed then released inside
+    /// Occurs when a button is pressed then released inside.
     case tap
     
-    /// Occurs when a button is double tapped
+    /// Occurs when a button is double tapped.
     case doubleTap
     
-    /// Occurs when a button is pressed down
+    /// Occurs when a button is pressed down.
     case press
     
-    /// Occurs when a button is released, inside or outside
+    /// Occurs when a button is released, inside or outside.
     case release
     
-    /// Occurs when a button is long pressed
+    /// Occurs when a button is long pressed.
     case longPress
     
-    /// Occurs repeatedly when a button is pressed and held
+    /// Occurs repeatedly when a button is pressed and held.
     case repeatPress
 }