+++ Sources/KeyboardKit/Proxy/UITextDocumentProxy+Delete.swift
@@ -20,7 +20,7 @@ public extension UITextDocumentProxy {
     /**
      Delete backwards a certain range.
      */
-    func deleteBackward(_ range: DeleteBackwardRange) {
+    func deleteBackward(range: DeleteBackwardRange) {
         guard let text = deleteBackwardText(for: range) else { return deleteBackward() }
         deleteBackward(times: text.count)
     }